Corrected get* functions to fix an error caused by na_if after a dplyr update
original
mutate_all(list(~na_if(.,"")))

new
mutate(across(everything(),~na_if(as.character(.),"")))

In a previous version of dplyr this line of code worked to change all "" (empty) cells of a df to NA, regardless if columns were character, vector or numeric:

 mutate_all(list(is.character, na_if, c("")))

After an update, this line is giving the error :
error in na_if():
! Can't convert y <character> to match type of x <integer>.

The core issue: you're trying to apply na_if() with a character ("") across all columns, but na_if() needs both x and y to be of the same type. It works fine for character columns, but will now throw an error on numeric columns because you're comparing a number to a character (""), which doesn't make sense.